Implement an unusually high number of CPU instructions...
Yeah I got bored of implementing CPU instructions one-by-one so now
there was a massive implementation cycle and each type of instruction
with a base available got implemented. Basic copy paste... I fear there
quite a few bugs because of that and it would important to be soon be
able to run blargg's tests.

Signed-off-by: Gyorgy Rethy <rethygy0410@gmail.com>